requirements.txt
# 深度学习核心库
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
opencv-python>=4.8.0

# 数据增强
albumentations>=1.3.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# 评估指标
scikit-learn>=1.3.0

# 进度条
tqdm>=4.65.0

# Kaggle API（用于下载数据）
kaggle>=1.5.16

# 模型融合（可选）
scipy>=1.10.0

# 其他工具
tensorboard>=2.13.0
